head: optionally indicate underrun of set limit
It is often convinient to detect whether head has in fact printed the
requested number of lines or if EOF was reached before reaching that
point. This patch adds the option --indicate-underrun, which makes
"head" exit with a status of 4 instead of 0 if no more data could be
read from its input.

This ability is useful when using head to chop a piped data stream
into slices and process it further:

datagenerator | while true; do
	head --indicate-underrun --bytes 2G | processdatachunk
	# check exit status of head to abort the loop on EOF
done

An example might be the creation of a tar archive while uploading it to
a broken FTP server in chunks of less than 2GB.